mhi_integration_engine/extractors/MSfourvenues/requirements.txt
functions-framework==3.0.0
google-cloud-secret-manager==2.19.0
google-cloud-storage==2.9.0
google-cloud-pubsub==2.22.0
pandas==2.2
pyarrow==16.1.0
requests==2.31.0
gcsfs==2024.6.1
jinja2==3.1.2